+++ ex_webrtc/CHANGELOG.md
@@ -25,3 +25,4 @@
 * Add voice activity detection [#33](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/33)
 * Fix track removed error in PeerConnectionHandler [#34](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/34)
 * Wait for PeerConnection to make a connection [#35](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/35)
+* Set track ready on first packet arrival [#36](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/36)